Add the ability to enable/disable a channel over WiFi
Currently the only way to enable or disable a channel's power is by
pressing a button. It can be useful to be able to do it via Wifi.

For the ability to work, the user needs to configure it via serial
command line. There, it can be:
* allowed from any IP address,
* allowed from IP address of the logging server,
* forbidden.

To enable or disable a channel's power via WiFi, an UDP packet
containing an 8 bytes long message (`enblChnN` or `dsblChnN`) must be
sent to the SmartPower3 device IP address, UDP port 3333.

This can be done for example with the socat utility:
```
echo -n enblChn0 | socat stdin udp-sendto:IP_ADDRESS:3333
```

The ability to toggle a channel's power can be useful for:
* Remote debugging. My SmartPower3 is at my office, powering devices
  that I am working on. I am developing U-Boot / Linux kernel drivers on
  these devices. Sometimes I need to disable the power for the device
  completely. Another plus is that when I am done working, I want to
  disable the power to the device completely, for safety.
* Automatic testing / CI. For example the U-Boot bootloader is actively
  developed, there are many changes being merged every week and these
  changes need to be automatically tested. Some devices need to be
  completely powered off/on to enter a mode that allows to upload
  firmware via serial. Having the ability to automatically power cycle
  a device is therefore helpful for automatic testing.